+++ lib/core/lib/model/update-instance-pool-details.ts
@@ -65,6 +65,7 @@ To use the instance pool with a regional subnet, provide a placement configurati
   "placementConfigurations"?: Array<model.UpdateInstancePoolPlacementConfigurationDetails>;
   /**
     * The number of instances that should be in the instance pool.
+* For a GMC-enabled resource pool, this is the number of GMC resources that should be in the pool.
 * <p>
 To determine whether capacity is available for a specific shape before you resize an instance pool,
 * use the {@link #createComputeCapacityReport(CreateComputeCapacityReportRequest) createComputeCapacityReport}